Splitte ut egen Context for oversiktsbildet
