+++ src/client/context/PersonContext.js
@@ -1,70 +1,14 @@
 import React, { createContext, useState } from 'react';
 import PropTypes from 'prop-types';
 import ErrorModal from '../components/ErrorModal';
-import { fetchPerson, fetchPersonoversikt, getPerson } from '../io/http';
+import { fetchPerson } from '../io/http';
 import { useSessionStorage } from '../hooks/useSessionStorage';
 
 export const PersonContext = createContext();
 
-const appendPersoninfo = behandling => {
-    return getPerson(behandling.originalSøknad.aktorId)
-        .then(response => ({
-            ...behandling,
-            personinfo: {
-                navn: response.data?.navn,
-                kjønn: response.data?.kjønn,
-                fnr: response.data?.fnr
-            }
-        }))
-        .catch(err => {
-            console.error('Feil ved henting av person.', err);
-            return behandling;
-        });
-};
-
 export const PersonProvider = ({ children }) => {
     const [error, setError] = useState(undefined);
     const [personTilBehandling, setPersonTilBehandling] = useSessionStorage('person');
-    const [personoversikt, setPersonoversikt] = useState([]);
-    const [isFetchingPersonoversikt, setIsFetchingPersonoversikt] = useState(false);
-    const [isFetchingPersoninfo, setIsFetchingPersoninfo] = useState(false);
-
-    const velgPersonFraOversikt = ({ aktørId }) => {
-        return hentPerson(aktørId);
-    };
-
-    const hentPersonoversikt = async () => {
-        setPersonTilBehandling(undefined);
-        const oversikt = await hentPersoner();
-        setPersonoversikt(oversikt);
-        setIsFetchingPersoninfo(true);
-        const oversiktWithPersoninfo = await Promise.all(
-            oversikt.map(behandling => appendPersoninfo(behandling))
-        ).finally(() => setIsFetchingPersoninfo(false));
-        if (oversiktWithPersoninfo.find(behandling => behandling.personinfo === undefined)) {
-            setError({ message: 'Kunne ikke hente navn for en eller flere saker. Viser aktørId' });
-        }
-        setPersonoversikt(oversiktWithPersoninfo);
-    };
-
-    const hentPersoner = () => {
-        setIsFetchingPersonoversikt(true);
-        return fetchPersonoversikt()
-            .then(response => response.data.personer)
-            .catch(err => {
-                setError({
-                    ...err,
-                    message:
-                        err.statusCode === 401
-                            ? 'Du må logge inn på nytt'
-                            : err.statusCode === 404
-                            ? 'Fant ingen behandlinger mellom i går og i dag.'
-                            : 'Kunne ikke hente behandlinger. Prøv igjen senere.'
-                });
-                return [];
-            })
-            .finally(() => setIsFetchingPersonoversikt(false));
-    };
 
     const hentPerson = value => {
         return fetchPerson(value)
@@ -87,13 +31,8 @@ export const PersonProvider = ({ children }) => {
     return (
         <PersonContext.Provider
             value={{
-                velgPersonFraOversikt,
-                personoversikt,
-                hentPersonoversikt,
                 personTilBehandling,
-                hentPerson,
-                isFetchingPersonoversikt,
-                isFetchingPersoninfo
+                hentPerson
             }}
         >
             {children}
